backend/requirements.txt
flask
flask-socketio
flask-sqlalchemy
flask-cors
eventlet
flask
flask-sqlalchemy
flask-jwt-extended
flask-bcrypt
flask-socketio
eventlet
